# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "minotari",
    "description": "",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.2.0"
  },
  "paths": {
    "/accounts/{name}/address": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves the Tari address for a specified account.",
        "description": "Returns the account's Tari address in Base58 format along with the emoji ID\nrepresentation. This address can be shared with others to receive payments.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Response\n\nReturns an [`AddressResponse`] object containing:\n- The address in Base58 encoding\n- The emoji ID representation\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Response\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\"\n}\n```",
        "operationId": "api_get_address",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve address for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account address retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/address_with_payment_id": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Creates a Tari address with an embedded payment ID for a specified account.",
        "description": "Generates a new address that includes a payment ID, which can be used to\nidentify specific transactions or invoices. When someone sends funds to\nthis address, the payment ID will be included in the transaction.\n\n# Path Parameters\n\n- `name`: The unique account name\n\n# Request Body\n\nSee [`CreatePaymentIdAddressRequest`] for the complete request schema.\n\n# Response\n\nReturns an [`AddressWithPaymentIdResponse`] object containing:\n- The address with embedded payment ID in Base58 encoding\n- The emoji ID representation\n- The original payment ID\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n- [`ApiError::InternalServerError`]: Failed to generate the address\n\n# Example Request\n\n```bash\ncurl -X POST http://localhost:8080/accounts/default/address_with_payment_id \\\n  -H \"Authorization: Bearer $MINOTARI_API_TOKEN\" \\\n  -H \"Content-Type: application/json\" \\\n  -d '{\"payment_id\": \"696e766f6963652d3132333435\"}'\n```\n\n# Example Response\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\",\n  \"payment_id_hex\": \"696e766f6963652d3132333435\"\n}\n```",
        "operationId": "api_create_address_with_payment_id",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to create address for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIdAddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Address with payment ID created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressWithPaymentIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{name}/balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves the current balance for a specified account.",
        "description": "Returns the account's available balance, pending incoming transactions,\nand locked funds. This endpoint is useful for displaying wallet status\nor checking available funds before initiating a transaction.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Response\n\nReturns an [`AccountBalance`] object containing:\n- Available (spendable) balance\n- Pending incoming balance\n- Locked balance (reserved for pending transactions)\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Response\n\n```json\n{\n  \"available\": 10000000,\n  \"pending_incoming\": 500000,\n  \"locked\": 200000\n}\n```",
        "operationId": "api_get_balance",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve balance for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBalance"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{name}/scan_status": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Retrieves the scan status for a specified account.",
        "description": "Returns information about the last scanned block, including the block height,\nblock hash, and the timestamp when the scan occurred. This is useful for\nmonitoring the wallet's synchronization progress with the blockchain.\n\n# Path Parameters\n\n- `name`: The unique account name to query\n\n# Response\n\nReturns a [`ScanStatusResponse`] object containing:\n- Last scanned block height\n- Last scanned block hash (hex encoded)\n- Timestamp when the block was scanned\n\n# Errors\n\n- [`ApiError::AccountNotFound`]: The specified account does not exist\nIf no blocks have been scanned yet, returns a response with default values (height 0, empty hash and timestamp).\n- [`ApiError::DbError`]: Database connection or query failure\n\n# Example Response\n\n```json\n{\n  \"last_scanned_height\": 12345,\n  \"last_scanned_block_hash\": \"abc123def456...\",\n  \"scanned_at\": \"2024-01-15 10:30:00\"\n}\n```",
        "operationId": "api_get_scan_status",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the account to retrieve scan status for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountBalance": {
        "type": "object",
        "required": [
          "total",
          "available",
          "locked",
          "unconfirmed",
          "immature"
        ],
        "properties": {
          "available": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "immature": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "locked": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "max_date": {
            "type": [
              "string",
              "null"
            ],
            "description": "The timestamp of the most recent transaction.\n\nThe string is in ISO 8601 format. Will be `None` if the\naccount has no transactions."
          },
          "max_height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The maximum blockchain height among all transactions for this account.\n\nWill be `None` if the account has no transactions."
          },
          "total": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "total_credits": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "total_debits": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          },
          "unconfirmed": {
            "type": "object",
            "properties": {
              "amount": {
                "type": "integer"
              }
            }
          }
        }
      },
      "AddressResponse": {
        "type": "object",
        "description": "API response type for an account address.\n\nContains the Tari address in Base58 format along with the emoji ID representation.\n\n# JSON Example\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\"\n}\n```",
        "required": [
          "address",
          "emoji_id"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "The Tari address in Base58 encoding"
          },
          "emoji_id": {
            "type": "string",
            "description": "The emoji representation of the address"
          }
        }
      },
      "AddressWithPaymentIdResponse": {
        "type": "object",
        "description": "API response type for an address with payment ID.\n\nContains the Tari address with embedded payment ID in Base58 format,\nalong with the emoji ID representation and the original payment ID in hex.\n\n# JSON Example\n\n```json\n{\n  \"address\": \"f4FxMqKAPDMqAjh6hTpCnLKfEu3MmS7NRu2YmKZPvZHc2K\",\n  \"emoji_id\": \"🎉🌟🚀...\",\n  \"payment_id_hex\": \"696e766f6963652d3132333435\"\n}\n```",
        "required": [
          "address",
          "emoji_id",
          "payment_id_hex"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "The Tari address with embedded payment ID in Base58 encoding"
          },
          "emoji_id": {
            "type": "string",
            "description": "The emoji representation of the address"
          },
          "payment_id_hex": {
            "type": "string",
            "description": "The payment ID that was embedded in the address (hex encoded)"
          }
        }
      },
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "description": "A general internal server error with a descriptive message.\n\nUsed for unexpected errors that don't fit other categories.\nReturns HTTP 500 Internal Server Error.",
            "required": [
              "InternalServerError"
            ],
            "properties": {
              "InternalServerError": {
                "type": "string",
                "description": "A general internal server error with a descriptive message.\n\nUsed for unexpected errors that don't fit other categories.\nReturns HTTP 500 Internal Server Error."
              }
            }
          },
          {
            "type": "object",
            "description": "A database operation failed.\n\nThis includes connection failures, query errors, and constraint\nviolations. Returns HTTP 500 Internal Server Error.\n\n# Example Causes\n\n- Database connection pool exhausted\n- SQL query syntax error\n- Foreign key constraint violation",
            "required": [
              "DbError"
            ],
            "properties": {
              "DbError": {
                "type": "string",
                "description": "A database operation failed.\n\nThis includes connection failures, query errors, and constraint\nviolations. Returns HTTP 500 Internal Server Error.\n\n# Example Causes\n\n- Database connection pool exhausted\n- SQL query syntax error\n- Foreign key constraint violation"
              }
            }
          },
          {
            "type": "object",
            "description": "The requested account was not found.\n\nThe contained string is the account name that was not found.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"Account 'nonexistent' not found\"\n}\n```",
            "required": [
              "AccountNotFound"
            ],
            "properties": {
              "AccountNotFound": {
                "type": "string",
                "description": "The requested account was not found.\n\nThe contained string is the account name that was not found.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"Account 'nonexistent' not found\"\n}\n```"
              }
            }
          },
          {
            "type": "object",
            "description": "A requested resource was not found.\n\nUsed for generic not found errors beyond just accounts.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"No blocks have been scanned yet\"\n}\n```",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "A requested resource was not found.\n\nUsed for generic not found errors beyond just accounts.\nReturns HTTP 404 Not Found.\n\n# Example\n\n```json\n{\n  \"error\": \"No blocks have been scanned yet\"\n}\n```"
              }
            }
          },
          {
            "type": "object",
            "description": "A bad request error for invalid input.\n\nUsed when the client sends invalid data in the request.\nReturns HTTP 400 Bad Request.\n\n# Example\n\n```json\n{\n  \"error\": \"Invalid hex in payment_id_hex: Invalid character\"\n}\n```",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "A bad request error for invalid input.\n\nUsed when the client sends invalid data in the request.\nReturns HTTP 400 Bad Request.\n\n# Example\n\n```json\n{\n  \"error\": \"Invalid hex in payment_id_hex: Invalid character\"\n}\n```"
              }
            }
          },
          {
            "type": "object",
            "description": "Failed to lock funds for a transaction.\n\nThis typically occurs when there are insufficient available funds\nor when UTXO selection fails. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance in the account\n- All UTXOs are already locked by other pending transactions\n- UTXO selection algorithm could not find suitable inputs",
            "required": [
              "FailedToLockFunds"
            ],
            "properties": {
              "FailedToLockFunds": {
                "type": "string",
                "description": "Failed to lock funds for a transaction.\n\nThis typically occurs when there are insufficient available funds\nor when UTXO selection fails. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance in the account\n- All UTXOs are already locked by other pending transactions\n- UTXO selection algorithm could not find suitable inputs"
              }
            }
          },
          {
            "type": "object",
            "description": "Failed to create an unsigned transaction.\n\nThis occurs during transaction construction after funds have been\nlocked. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Invalid recipient address format\n- Transaction size exceeds limits\n- Cryptographic operation failure",
            "required": [
              "FailedCreateUnsignedTx"
            ],
            "properties": {
              "FailedCreateUnsignedTx": {
                "type": "string",
                "description": "Failed to create an unsigned transaction.\n\nThis occurs during transaction construction after funds have been\nlocked. Returns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Invalid recipient address format\n- Transaction size exceeds limits\n- Cryptographic operation failure"
              }
            }
          },
          {
            "type": "object",
            "description": "Failed to build or broadcast a burn transaction.\n\nThis occurs during burn transaction construction or network submission.\nReturns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance\n- Invalid claim public key\n- Base node rejected the transaction",
            "required": [
              "FailedToBurnFunds"
            ],
            "properties": {
              "FailedToBurnFunds": {
                "type": "string",
                "description": "Failed to build or broadcast a burn transaction.\n\nThis occurs during burn transaction construction or network submission.\nReturns HTTP 500 Internal Server Error.\n\n# Common Causes\n\n- Insufficient balance\n- Invalid claim public key\n- Base node rejected the transaction"
              }
            }
          }
        ],
        "description": "Represents all possible errors returned by the REST API.\n\nEach variant corresponds to a specific error condition that can occur\nduring API request processing. The error type automatically converts\nto an appropriate HTTP response with a JSON error body.\n\n# Error Handling Pattern\n\nAPI handlers typically use the `?` operator with this error type:\n\n```rust,ignore\npub async fn handler() -> Result<Json<Data>, ApiError> {\n    let data = fetch_data().await?; // Errors automatically convert to ApiError\n    Ok(Json(data))\n}\n```\n\n# Serialization\n\nWhen serialized to JSON for API responses, errors produce:\n\n```json\n{\n  \"error\": \"Error message here\"\n}\n```"
      },
...
      "CreatePaymentIdAddressRequest": {
        "type": "object",
        "description": "Request body for creating an address with a payment ID.\n\n# JSON Example\n\n```json\n{\n  \"payment_id_hex\": \"696e766f6963652d3132333435\"\n}\n```",
        "required": [
          "payment_id_hex"
        ],
        "properties": {
          "payment_id_hex": {
            "type": "string",
            "description": "The payment ID to embed in the address, hex encoded.\n\nThis should be a hex-encoded byte string (e.g., \"696e766f6963652d3132333435\" for \"invoice-12345\").\nMaximum length is 256 bytes when decoded."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Wallet API token, sent as 'X-API-Key: <token>'"
      },
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "Wallet API token, sent as 'Authorization: Bearer <token>'"
      }
    }
  },
  "security": [
    {
      "bearer_token": []
    },
    {
      "api_key": []
    }
  ],
...
}